Use CacheWindow in Pager
Use CacheWindow in Pager. Pager will now use Cache Window to prefetch and cache pages, instead of a single page in the direction of the scroll. This should work in addition to beyond bounds page count.

Relnote: Update Pager's prefetch strategy to use Cache Window instead of the default 1 item in the direction of the scroll. Now initial prefetching is enabled by default in Pager and prefetching will be based on the size of the window worth 1 viewport, that is, we will try to fill 1 whole view port with prefetched items. Items will be kept around longer as well.

Test: Added new tests.

Fixes: 315280218
Fixes: 292136289
Change-Id: I4d45ece5bab78dc241618d29d3701e5463789aa6